pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nash-llm-server"
version = "0.1.11"
description = "Nash LLM Server"
authors = ["Nash Team <team@nash.run>"]
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11"
requests = ">=2.32.3,<3.0.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
mcp = {version = ">=1.3.0,<2.0.0", extras = ["cli"]}
litellm = "^1.63.11"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.0.0"

[tool.poetry.scripts]
llm_server = "app.server:main"
client_example = "client_example:main"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120
